SplashScreen rewrite
