+++ feature/auth/src/commonMain/kotlin/com/hulkdx/findprofessional/feature/authentication/splash/SplashViewModel.kt
@@ -2,16 +2,13 @@ package com.hulkdx.findprofessional.feature.authentication.splash
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.hulkdx.findprofessional.common.feature.authentication.login.LoginUseCase
 import kotlinx.coroutines.launch
 
 class SplashViewModel(
-    useCase: LoginUseCase,
 ) : ViewModel() {
 
     init {
         viewModelScope.launch {
-            useCase.onSplashScreenLoaded()
         }
     }
